combine toasts into single component
